+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 6/15/2023
 - Major: Moved `Discord`, `Telegram` to controllers folder
 - Major: Added `Account` class to handle account related functions
+- Major: Added template for classes
 
 ## 6/12/2023
 - Minor: Added `skipCheck` option to bypass `account.fired` check and will always send a notification.